requirements.txt
asgiref==3.4.1
brotlipy==0.7.0
certifi==2021.5.30
cffi
chardet
colorama==0.4.4
conda==4.10.3
conda-package-handling
coreapi==2.3.3
coreschema==0.0.4
cryptography
cycler==0.11.0
Django==3.2.9
django-admin-list-filter-dropdown==1.0.3
django-cors-headers==3.10.0
django-extensions==3.1.5
django-filter==21.1
django-json-widget==1.1.1
django-safedelete==1.1.0
djangorestframework==3.12.4
djangorestframework-datatables==0.6.0
djangorestframework-simplejwt==5.0.0
drf-flex-fields==0.9.5
drf-yasg==1.20.0
et-xmlfile==1.1.0
future==0.18.2
fuzzywuzzy==0.18.0
gunicorn==20.1.0
idna
importlib-metadata==4.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
kiwisolver==1.3.2
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.4.3
menuinst==1.4.16
numpy==1.21.3
openpyxl==3.0.9
packaging==21.2
pandas==1.3.4
Pillow==8.4.0
psycopg2==2.9.2
pycosat==0.6.3
pycparser
PyJWT==2.3.0
pyodbc==4.0.32
pyOpenSSL
pyparsing==2.4.7
PySocks
python-dateutil==2.8.2
pytz==2021.3
pywin32==228
requests
ruamel-yaml-conda
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
scipy==1.7.2
seaborn==0.11.2
six
sqlparse==0.4.2
tqdm
uritemplate==4.1.1
urllib3
win-inet-pton
wincertstore==0.2
zipp==3.7.0
